// src/app/models/pet.model.ts
export interface Pet {
    publication_id: number;
    user_id: number;
    pet_filename: string;
    pet_name: string;
    pet_species: string;
    pet_sex: string;
    status: string;
    pet_last_location: string;
    pet_breed: string;
    user_filename: string;
    user_first_name: string;
}
  